Fix SMB Scanner thread could run in the background
After a first smb scan, when it is completed or cancelled, the next scans started from the same dialog permit to select a server from displayed results while the scan is running.
This caused the scan to continue running in the background until it was completed. Further scans could be even started while the first scan was not completed causing a significant network hammering.
Now, the scan results can no longer be selected until a scan thread is cancelled or completed